requirements.txt
tensorflow-gpu~=2.2.0
numpy~=1.17.0
six~=1.12.0
scipy~=1.4.1
pillow~=6.1.0
matplotlib~=3.1.0
sklearn
scikit-learn~=0.21.3
opencv-python~=4.1.1.26
h5py~=2.10.0
pyyaml~=5.3
imgaug~=0.4.0
pandas~=1.0.2
future~=0.18.2
munch~=2.5.0
scikit-image~=0.15.0
tensorflow_addons